requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
openai==2.30.0
openai-agents==0.13.2
python-dotenv==1.2.2
pydantic==2.12.5
discord.py==2.7.1
aiohttp==3.13.4

# dev tools
ruff==0.15.8
mypy==1.19.1
